requirements.txt
asyncpg==0.30.0
async_lru==2.0.5
fastapi==0.115.12
gunicorn==23.0.0
numpy==1.26.4
pandas==2.2.3
pydantic==2.11.3
python-dotenv==1.1.0
Requests==2.32.3
SQLAlchemy==2.0.36
uvicorn==0.34.0
